--- src/main/kotlin/com/teamsparta/tikitaka/domain/match/dto/dto.kt
@@ -1,4 +1,0 @@
-package com.teamsparta.tikitaka.domain.match.dto
-
-class dto {
-}
