+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/repository/MatchRepository.kt
@@ -1,4 +1,10 @@
 package com.teamsparta.tikitaka.domain.match.repository
 
-interface MatchRepository {
+import com.teamsparta.tikitaka.domain.match.model.Match
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface MatchRepository : JpaRepository<Match, Long> {
+    fun findByDeletedAtIsNull(): List<Match>
 }
