--- src/main/kotlin/com/teamsparta/tikitaka/domain/matching/dto/UpdateMatchRequest.kt
@@ -1,11 +1,0 @@
-package com.teamsparta.tikitaka.domain.matching.dto
-
-import java.time.LocalDateTime
-
-class UpdateMatchRequest (
-    val title:String,
-    val matchDate : LocalDateTime,
-    val location : String,
-    val content : String,
-    val matchStatus : Boolean,
-)
